update short answer question answer content
